First Stable Build
This scrapbook of code was cobbled together from the work of other
people who actually know what they are doing, and my only real
contribution was the hours I spent wading through forum threads
collecting examples, before bolting this Frankenstein together.  So I
take no credit for any of it, and provide no guarantees whatsoever that
it wont brick your system the first time you run it. I also hope that
anyone I’ve missed in the credits, forgives the accidental oversight, as
some details get lost in the process of all that cutting and pasting.