adjust second interval day rollover
When using second level rtc interrupts, my test units only got to the
midnight rollover, and then failed to set the next alarm. I suspect this
is because I used the 8sec sleeps to wait for the rollover, so I just
replaced it with delay statements. I only use sub minute alarms for
debugging, so  does not affect normal operations with RTC alarms longer
than 1 minute.